v0breaking: converted classes to plain object
- The nodes are no longer classes but simple objects. Readonly has been removed
from all the properties of their interfaces and the parser result is not longer
"sealed". They should still not be mutated, but if you really want to you can.

- Nodes no longer have a parent property to avoid circular references. Instead
you can generate a parents map from a parser result to use getParent on if needed.

- To check a node or token is of a certain type, they now all have a `type` property.

- To check if an object is a node or a token,
you can use the new `isNode` and `isToken` utility functions.

- The non-node normalized Condition and Expression classes
are now the NormalizedCondition and NormalizedExpression interfaces.
Their object will return false if passed to isNode/Token, but they
have an AST_TYPE to be able to tell them apart.

- All node creation functions have been moved to /ast/*, /ast/classes has been removed.

- Several internal helpers have been moved, there is now an internals export
(just in case), it's not really meant to be used.