v0fix: fixed prefixApplier being called when groups are negated
A negation is not a prefix and it shouldn't have been getting called. It was
semi-working because the default prefixApplier just joins it and a blank prefix
was being passed, so "" + value = value. But when the prefixApplier used dot
separation, negating like `!(a)` was creating keys that looked like `.a`.