v0breaking: added support for typescript's "erasableSyntaxOnly"
This is a breaking change because this required replacing all enums with plain
objects. This requires some refactoring if using any of the enums as types (if
there is an enum `SOME_ENUM`, it's type is now exported as `SomeEnum` (where
possible) and values should be referenced like `typeof SOME_ENUM.KEY`). Also
some enums/types had to be renamed to avoid conflicts with existing types
(mostly all the lexer token types).